requirements.txt
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.2
scipy==1.10.1
